+++ test/test_prune.mli
@@ -1,0 +1,4 @@
+(** Unit tests for [Prune]. *)
+
+val suite : string * unit Alcotest.test_case list
+(** Alcotest suite for the [Prune] module. *)